+++ divmod/shift.go
@@ -2,7 +2,7 @@ package divmod
 
 import "github.com/borisskert/go-biginteger/digits"
 
-func ShiftRightUint64Array(a []uint64, n uint64) []uint64 {
+func ShiftRightUint64Array(a []uint64, n uint) []uint64 {
 	shifted := digits.Wrap(a).RightShiftBits(n)
 	shifted.NormalizeInPlace()
